requirements.txt
pandas>=1.3.0
spacy>=3.0.0
together
python-dotenv>=0.19.0
streamlit